// src/app/models/userloan.ts
export class UserLoan {
  uId: string;
  loanType: string;
  loanAmount: string;
  applyDate: string;
  rateOfInterest: number;
  loanDuration: string;
  courseFee: string;
  courseName: string;
  fatherName: string;
  fatherOcc: string;
  fatherExp: number;
  rationCard: string;
  fatherInc: string;
  annualIncome: string;
  companyName: string;
  desig: string;
  totalExp: number;
  currExp: number;
}